requirements.txt
discord.py
requests
matplotlib